Add integraiton test rearrage python scripts
