+++ package.json
@@ -1,6 +1,13 @@
 {
+  "name": "test app",
+  "version": "1.0.0",
+  "scripts": {
+    "start": "concurrently \"node app.js\" \"node cron.js\""
+  },
   "dependencies": {
+    "concurrently": "^7.3.0",
     "croner": "^4.2.0",
-    "dotenv": "^16.0.1"
+    "dotenv": "^16.0.1",
+    "express": "^4.18.1"
   }
 }
