+++ requirements.txt
@@ -1,6 +1,6 @@
 requests~=2.27.1
 pytest==6.2.2
-botocore==1.20.95
+botocore==1.27.55
 boto3==1.24.55
 PyJWT==2.4.0
 cryptography==37.0.2
